\name{internal}
\alias{internal}
\alias{buildTrainingSet}
\alias{buildTrainingSet.signaltrans}
\alias{code_test}
\alias{code_train}
\alias{getInterProDomains}
\alias{getKEGGHierarchy}
\alias{loss}
\alias{predict.gene2pathway}
\alias{struct_predict}
\alias{struct_train}
\alias{svmlearn}
\alias{svmpredict}
\alias{get.element.relations.by.pathway}
\alias{KEGG2Entrez}
\alias{getComponents.internal}
\alias{Entrez2ORF.internal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{  internal functions  }
\description{
  internal functions: do not call these functions directly. 
}
\usage{
	various
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	various
}
\value{
 	various
}
\author{ Holger Froehlich }
\keyword{file}

